Added Caesar cipher.
There is no way to access the Vigenere cipher from the main program. The TODO
file has been accordingly updated. Also, note that there is a LOT of repetitive
code. However, it is somewhat difficult to abstract away because it would
require currying to do cleanly, or a different approach.